ai/requirements.txt
# AI Service Dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
onnxruntime==1.18.0
tokenizers==0.15.2
numpy>=1.26.4,<2.0
